refactor(analytics): fix main_test.go build error
